+++ ricaun.RevitTest.Shared/TestRequest.cs
@@ -15,7 +15,8 @@ namespace ricaun.RevitTest.Shared
         private void NotifyPropertyChanged([CallerMemberName] string propertyName = "") =>
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         #endregion
-        public string TestFilter { get; set; }
+        public string Info { get; set; }
+        public string[] TestFilters { get; set; }
         public string TestPathFile { get; set; }
         public override string ToString()
         {