+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] / 2023-04-17
+### Features
+[x] Update `TestAdapter` to work with '.' in the TestName. ('TestCaseUtils') 
+[x] Update `RevitTest` to work with ',' and '\\' in the TestName.
+### Application
+[x] Update `Tests` without Split(',')
+### Console
+[x] Update Command `Test` to multiple parameters using `IEnumerable`.
+### Shared
+[x] Add Info to `TestRequest`
+[x] Update `TestFilter` to `TestFilters` with `string[]`
+### TestAdapter
+[x] Update `ProcessStart` to work with `IEnumerable` arguments, join with ' '.
+[x] Update `RevitTestProcessStart` to work filters `string[]`. 
+
 ## [1.0.2] / 2023-04-15
 ### Features
 [x] Update `ricaun.NUnit` to 1.2.7
@@ -82,6 +97,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.3]: ../../compare/1.0.2...1.0.3
 [1.0.2]: ../../compare/1.0.1...1.0.2
 [1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/1.0.0